remove supporting functions to utils folder refactor the "images are the same" test to a series of assertions PR feedback
